# frontend/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "InvoiceFlow API",
    "description": "E-invoice processing platform with AI extraction and global compliance (PEPPOL, ZATCA, Brazil NFe, India IRP, Mexico CFDI, CTC reporting, and post-audit archival).",
    "version": "0.2.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Local development" },
    { "url": "/api/v1", "description": "Production (Nginx proxy)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "InvoiceStatus": {
        "type": "string",
        "enum": ["received", "processing", "extracted", "reviewing", "approved", "compliant", "transmitted", "failed", "rejected"]
      },
...
      "TokenRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email", "password"]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "default": "bearer" },
          "user_id": { "type": "string" },
          "tenant_id": { "type": "string" }
        },
        "required": ["access_token", "user_id", "tenant_id"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "full_name": { "type": "string", "nullable": true },
          "tenant_name": { "type": "string" },
          "tenant_slug": { "type": "string" },
          "country_code": { "type": "string", "default": "US" }
        },
        "required": ["email", "password", "tenant_name", "tenant_slug"]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "full_name": { "type": "string", "nullable": true },
          "role": { "type": "string" },
          "tenant_id": { "type": "string", "format": "uuid" }
        },
        "required": ["id", "email", "role", "tenant_id"]
      },
      "FileUploadResponse": {
        "type": "object",
        "properties": {
          "invoice_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Null when processing is async; use task_id for tracking" },
          "status": { "$ref": "#/components/schemas/InvoiceStatus" },
          "filename": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["status", "filename", "message"]
      },
...
      "WebhookPayload": {
        "type": "object",
        "properties": {
          "filename": { "type": "string" },
          "file_content_b64": { "type": "string", "description": "Base64-encoded file content" },
          "mime_type": { "type": "string", "default": "application/pdf" },
          "source_reference": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "required": ["filename", "file_content_b64"]
      },
...
      "TenantSettings": {
        "type": "object",
        "properties": {
          "settings": {
            "type": "object",
            "properties": {
              "tenant_name": { "type": "string" },
              "country_code": { "type": "string" },
              "default_currency": { "type": "string" },
              "default_compliance_model": { "type": "string" },
              "tax_id": { "type": "string", "nullable": true },
              "registration_number": { "type": "string", "nullable": true },
              "custom_settings": { "type": "object", "additionalProperties": true }
            }
          }
        },
        "required": ["settings"]
      },
      "CountryConfig": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "name": { "type": "string" },
          "model": { "type": "string" },
          "system": { "type": "string" }
        },
        "required": ["code", "name", "model", "system"]
      },
      "CountryConfigList": {
        "type": "object",
        "properties": {
          "countries": { "type": "array", "items": { "$ref": "#/components/schemas/CountryConfig" } }
        },
        "required": ["countries"]
      },
...
      "WebhookCreate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "url": { "type": "string", "format": "uri" },
          "secret": { "type": "string", "nullable": true },
          "events": { "type": "array", "items": { "type": "string" } },
          "content_type": { "type": "string", "default": "application/json" },
          "timeout_seconds": { "type": "integer", "default": 30 },
          "max_retries": { "type": "integer", "default": 3 },
          "active": { "type": "boolean", "default": true }
        },
        "required": ["name", "url", "events"]
      },
      "WebhookUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "nullable": true },
          "url": { "type": "string", "format": "uri", "nullable": true },
          "secret": { "type": "string", "nullable": true },
          "events": { "type": "array", "items": { "type": "string" }, "nullable": true },
          "content_type": { "type": "string", "nullable": true },
          "timeout_seconds": { "type": "integer", "nullable": true },
          "max_retries": { "type": "integer", "nullable": true },
          "active": { "type": "boolean", "nullable": true }
        }
      },
      "WebhookResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "url": { "type": "string" },
          "active": { "type": "boolean" },
          "events": { "type": "array", "items": { "type": "string" } },
          "content_type": { "type": "string" },
          "timeout_seconds": { "type": "integer" },
          "max_retries": { "type": "integer" },
          "last_triggered_at": { "type": "string", "format": "date-time", "nullable": true },
          "last_status_code": { "type": "integer", "nullable": true },
          "success_count": { "type": "integer" },
          "failure_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "tenant_id", "name", "url", "active", "events", "content_type", "timeout_seconds", "max_retries", "success_count", "failure_count", "created_at", "updated_at"]
      },
      "WebhookTestResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "status_code": { "type": "integer", "nullable": true },
          "message": { "type": "string" },
          "response_time_ms": { "type": "number" }
        },
        "required": ["success", "message", "response_time_ms"]
      },
      "WebhookEventList": {
        "type": "array",
        "items": { "type": "string" }
      },
...
      "ToggleResponse": {
        "type": "object",
        "properties": {
          "active": { "type": "boolean" }
        },
        "required": ["active"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string" }
        },
        "required": ["detail"]
      }
    }
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login and receive JWT token",
        "operationId": "login",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenRequest" } } } },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Register a new user and tenant",
        "operationId": "register",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } } },
        "responses": {
          "201": { "description": "Registration successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "409": { "description": "Email already registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["auth"],
        "summary": "Get current user info",
        "operationId": "getCurrentUser",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Current user info", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/v1/ingestion/webhook": {
      "post": {
        "tags": ["ingestion"],
        "summary": "Receive invoice via webhook",
        "operationId": "webhookReceive",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookPayload" } } } },
        "responses": {
          "200": { "description": "Webhook received", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileUploadResponse" } } } }
        }
      }
    },
...
    "/api/v1/settings/": {
      "get": {
        "tags": ["settings"],
        "summary": "Get tenant settings",
        "operationId": "getSettings",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Tenant settings", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TenantSettings" } } } }
        }
      },
      "put": {
        "tags": ["settings"],
        "summary": "Update tenant settings",
        "operationId": "updateSettings",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
        "responses": {
          "200": { "description": "Settings updated", "content": { "application/json": { "schema": { "type": "object", "properties": { "updated": { "type": "boolean" } }, "required": ["updated"] } } } }
        }
      }
    },
    "/api/v1/settings/countries": {
      "get": {
        "tags": ["settings"],
        "summary": "List available country compliance configurations",
        "operationId": "listCountryConfigs",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Country configs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CountryConfigList" } } } }
        }
      }
    },
...
    "/api/v1/webhooks/events": {
      "get": {
        "tags": ["webhooks"],
        "summary": "List supported webhook event types",
        "operationId": "listWebhookEvents",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Supported events", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookEventList" } } } }
        }
      }
    },
    "/api/v1/webhooks/": {
      "get": {
        "tags": ["webhooks"],
        "summary": "List webhook configurations",
        "operationId": "listWebhooks",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Webhook configs", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookResponse" } } } } }
        }
      },
      "post": {
        "tags": ["webhooks"],
        "summary": "Create webhook configuration",
        "operationId": "createWebhook",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookCreate" } } } },
        "responses": {
          "201": { "description": "Webhook created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookResponse" } } } }
        }
      }
    },
    "/api/v1/webhooks/{webhook_id}": {
      "get": {
        "tags": ["webhooks"],
        "summary": "Get webhook configuration",
        "operationId": "getWebhook",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "webhook_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Webhook details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookResponse" } } } }
        }
      },
      "put": {
        "tags": ["webhooks"],
        "summary": "Update webhook configuration",
        "operationId": "updateWebhook",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "webhook_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookUpdate" } } } },
        "responses": {
          "200": { "description": "Webhook updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookResponse" } } } }
        }
      },
      "delete": {
        "tags": ["webhooks"],
        "summary": "Delete webhook",
        "operationId": "deleteWebhook",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "webhook_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "204": { "description": "Webhook deleted" } }
      }
    },
    "/api/v1/webhooks/{webhook_id}/test": {
      "post": {
        "tags": ["webhooks"],
        "summary": "Send test event to webhook",
        "operationId": "testWebhook",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "webhook_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Test result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookTestResponse" } } } }
        }
      }
    },
    "/api/v1/webhooks/{webhook_id}/toggle": {
      "post": {
        "tags": ["webhooks"],
        "summary": "Toggle webhook active/inactive",
        "operationId": "toggleWebhook",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "webhook_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Toggle result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ToggleResponse" } } } }
        }
      }
    }
  },
  "security": [{ "BearerAuth": [] }]
}